added process-file-test.
Need to more clearly define onsite-block vs offsite-block and have logic names follow those definitions.

An onsite block represents a file that will be backed up and is stored in the DB of the offsite-cli.

An offsite block represents a point in time for a file or a sub-block of a file that is created in offsite-cli but ultimately is stored in offsite-node